@jorgecarleitao 2021-04-21
The check build is failing for some reason. Any idea why? Maybe a bump of the stable rust?