@Dandandan 2021-04-21
Not sure... Maybe a recent upgrade to the crate that has a compile error?