@tustvold 2023-01-04
FYI this is being reverted in https://github.com/apache/arrow-rs/pull/9 as it is causing pain, and doesn't have a well defined maintenance story. The recommendation continues to be that people compile with the relevant features enabled